+++ catapult-sdk/src/plugins/mosaic.js
@@ -31,15 +31,15 @@ const mosaicPlugin = {
 		builder.addTransactionSupport(EntityType.mosaicDefinition, {
 			id: ModelType.uint64HexIdentifier,
 			duration: ModelType.uint64,
-			nonce: ModelType.int,
-			flags: ModelType.int,
-			divisibility: ModelType.int
+			nonce: ModelType.uint32,
+			flags: ModelType.uint8,
+			divisibility: ModelType.uint8
 		});
 
 		builder.addTransactionSupport(EntityType.mosaicSupplyChange, {
 			mosaicId: ModelType.uint64HexIdentifier,
 			delta: ModelType.uint64,
-			action: ModelType.int
+			action: ModelType.uint8
 		});
 
 		builder.addSchema('mosaicDescriptor', {
@@ -53,8 +53,8 @@ const mosaicPlugin = {
 			startHeight: ModelType.uint64,
 			ownerAddress: ModelType.binary,
 			revision: ModelType.int,
-			flags: ModelType.int,
-			divisibility: ModelType.int,
+			flags: ModelType.uint8,
+			divisibility: ModelType.uint8,
 			duration: ModelType.uint64
 		});
 	},
